Puts pagination at bottom of page and changes placeholder
