--- components/pages/library/library-filters/library-filters-constants.js
@@ -1,22 +1,0 @@
-import React from 'react';
-
-export const TABLE_SIZE_VALUES = [
-  {
-    label: '5',
-    value: 5
-  },
-  {
-    label: '10',
-    value: 10
-  },
-  {
-    label: '15',
-    value: 15
-  },
-  {
-    label: '20',
-    value: 20
-  },
-];
-
-export default { TABLE_SIZE_VALUES };
