+++ components/pages/library/library-table/library-table-styles.scss
@@ -6,9 +6,15 @@
     margin: 25px 0 0;
   }
 
+  .pagination-filter {
+    display: flex;
+    justify-content: flex-end;
+    margin: 35px 0 0;
+  }
+
   .paginator-container {
     display: flex;
     justify-content: center;
-    margin: 55px 0 0;
+    margin: 25px 0 0;
   }
 }
